requirements.txt
python-dotenv
Requests
google-generativeai
Django
django_multiselectfield
